classes/class-07-mcp-protocol/demos/ai-agent-application/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
attrs==25.4.0
backoff==2.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
diff-match-patch==20241021
distro==1.9.0
faker==37.12.0
fastapi==0.121.0
fastuuid==0.14.0
filelock==3.20.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
guardrails-ai==0.6.7
guardrails-api-client==0.4.0
guardrails-hub-types==0.0.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.1.2
idna==3.11
importlib-metadata==8.7.0
isoduration==20.11.0
jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-core==0.3.79
langchain-mcp-adapters==0.1.12
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langfuse==3.9.1
langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.41
lark==1.3.1
lazy-imports==1.1.0
litellm==1.79.1
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.20.0
mdurl==0.1.2
multidict==6.7.0
openai==1.109.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pip==25.3
propcache==0.4.1
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.11.0
pydash==8.0.5
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
rstr==3.2.2
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
sse-starlette==3.0.3
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.15.4
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
webcolors==25.10.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
